requirements.txt
asgiref==3.9.1
Django==5.2.4
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.10
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0